+++ examples/date.rs
@@ -62,9 +62,9 @@ pub struct DateTime {
 
 fn two_digits<Input>() -> impl Parser<Input, Output = i32>
 where
-    Input: Stream<Item = char>,
+    Input: Stream<Token = char>,
     // Necessary due to rust-lang/rust#24159
-    Input::Error: ParseError<Input::Item, Input::Range, Input::Position>,
+    Input::Error: ParseError<Input::Token, Input::Range, Input::Position>,
 {
     (digit(), digit()).map(|(x, y): (char, char)| {
         let x = x.to_digit(10).expect("digit");
@@ -80,8 +80,8 @@ where
 /// Z
 fn time_zone<Input>() -> impl Parser<Input, Output = i32>
 where
-    Input: Stream<Item = char>,
-    Input::Error: ParseError<Input::Item, Input::Range, Input::Position>,
+    Input: Stream<Token = char>,
+    Input::Error: ParseError<Input::Token, Input::Range, Input::Position>,
 {
     let utc = char('Z').map(|_| 0);
     let offset = (
@@ -105,8 +105,8 @@ where
 /// 2010-01-30
 fn date<Input>() -> impl Parser<Input, Output = Date>
 where
-    Input: Stream<Item = char>,
-    Input::Error: ParseError<Input::Item, Input::Range, Input::Position>,
+    Input: Stream<Token = char>,
+    Input::Error: ParseError<Input::Token, Input::Range, Input::Position>,
 {
     (
         many::<String, _, _>(digit()),
@@ -129,8 +129,8 @@ where
 /// 12:30:02
 fn time<Input>() -> impl Parser<Input, Output = Time>
 where
-    Input: Stream<Item = char>,
-    Input::Error: ParseError<Input::Item, Input::Range, Input::Position>,
+    Input: Stream<Token = char>,
+    Input::Error: ParseError<Input::Token, Input::Range, Input::Position>,
 {
     (
         two_digits(),
@@ -155,8 +155,8 @@ where
 /// 2015-08-02T18:54:42+02
 fn date_time<Input>() -> impl Parser<Input, Output = DateTime>
 where
-    Input: Stream<Item = char>,
-    Input::Error: ParseError<Input::Item, Input::Range, Input::Position>,
+    Input: Stream<Token = char>,
+    Input::Error: ParseError<Input::Token, Input::Range, Input::Position>,
 {
     (date(), char('T'), time()).map(|(date, _, time)| DateTime {
         date: date,